+++ src/php/Compiler/Domain/Parser/ExpressionParser/ListParser.php
@@ -4,9 +4,9 @@ declare(strict_types=1);
 
 namespace Phel\Compiler\Domain\Parser\ExpressionParser;
 
-use Phel\Compiler\Application\Parser;
 use Phel\Compiler\Domain\Lexer\TokenStream;
 use Phel\Compiler\Domain\Parser\Exceptions\UnfinishedParserException;
+use Phel\Compiler\Domain\Parser\ParserInterface;
 use Phel\Shared\Exceptions\ErrorCode;
 use Phel\Shared\Parser\Node\ListNode;
 use Phel\Shared\Parser\Node\ReaderCondSplicingNode;
@@ -22,7 +22,7 @@ final readonly class ListParser
         Token::T_CLOSE_BRACE => '}',
     ];
 
-    public function __construct(private Parser $parser) {}
+    public function __construct(private ParserInterface $parser) {}
 
     /**
      * @throws UnfinishedParserException