+++ src/php/Compiler/Domain/Parser/ExpressionParser/QuoteParser.php
@@ -4,13 +4,13 @@ declare(strict_types=1);
 
 namespace Phel\Compiler\Domain\Parser\ExpressionParser;
 
-use Phel\Compiler\Application\Parser;
 use Phel\Compiler\Domain\Lexer\TokenStream;
+use Phel\Compiler\Domain\Parser\ParserInterface;
 use Phel\Shared\Parser\Node\QuoteNode;
 
 final readonly class QuoteParser
 {
-    public function __construct(private Parser $parser) {}
+    public function __construct(private ParserInterface $parser) {}
 
     public function parse(TokenStream $tokenStream, int $tokenType): QuoteNode
     {