+++ src/php/Compiler/Domain/Reader/ExpressionReader/MapReader.php
@@ -5,15 +5,15 @@ declare(strict_types=1);
 namespace Phel\Compiler\Domain\Reader\ExpressionReader;
 
 use Phel;
-use Phel\Compiler\Application\Reader;
 use Phel\Compiler\Domain\Reader\Exceptions\ReaderException;
+use Phel\Compiler\Domain\Reader\ReaderInterface;
 use Phel\Lang\Collections\Map\PersistentMapInterface;
 use Phel\Shared\Parser\Node\ListNode;
 use Phel\Shared\Parser\Node\NodeInterface;
 
 final readonly class MapReader
 {
-    public function __construct(private Reader $reader) {}
+    public function __construct(private ReaderInterface $reader) {}
 
     /**
      * @return PersistentMapInterface<mixed, mixed>