+++ src/php/Compiler/Domain/Reader/ExpressionReader/MetaReader.php
@@ -5,8 +5,8 @@ declare(strict_types=1);
 namespace Phel\Compiler\Domain\Reader\ExpressionReader;
 
 use Phel;
-use Phel\Compiler\Application\Reader;
 use Phel\Compiler\Domain\Reader\Exceptions\ReaderException;
+use Phel\Compiler\Domain\Reader\ReaderInterface;
 use Phel\Lang\Collections\Map\PersistentMapInterface;
 use Phel\Lang\Keyword;
 use Phel\Lang\MetaInterface;
@@ -19,7 +19,7 @@ use function is_string;
 
 final readonly class MetaReader
 {
-    public function __construct(private Reader $reader) {}
+    public function __construct(private ReaderInterface $reader) {}
 
     /**
      * @throws ReaderException