+++ src/php/Compiler/Domain/Reader/ExpressionReader/QuoasiquoteReader.php
@@ -4,10 +4,10 @@ declare(strict_types=1);
 
 namespace Phel\Compiler\Domain\Reader\ExpressionReader;
 
-use Phel\Compiler\Application\Reader;
 use Phel\Compiler\Domain\Reader\Exceptions\ReaderException;
 use Phel\Compiler\Domain\Reader\Exceptions\SpliceNotInListException;
 use Phel\Compiler\Domain\Reader\QuasiquoteTransformerInterface;
+use Phel\Compiler\Domain\Reader\ReaderInterface;
 use Phel\Lang\TypeInterface;
 use Phel\Shared\Parser\Node\NodeInterface;
 use Phel\Shared\Parser\Node\QuoteNode;
@@ -15,7 +15,7 @@ use Phel\Shared\Parser\Node\QuoteNode;
 final readonly class QuoasiquoteReader
 {
     public function __construct(
-        private Reader $reader,
+        private ReaderInterface $reader,
         private QuasiquoteTransformerInterface $quasiquoteTransformer,
     ) {}